+++ R/convert_d2etasq.R
@@ -1,7 +1,7 @@
-#' @title Convert effect size d into f
-#' @name esc_d2f
+#' @title Convert effect size d into Eta Squared
+#' @name convert_d2etasq
 #'
-#' @description Compute effect size \code{f} from effect size \code{d}.
+#' @description Compute effect size Eta Squared from effect size \code{d}.
 #'
 #' @param d The effect size \code{d}.
 #' @param se The standard error of \code{d}. One of \code{se} or \code{v}
@@ -11,7 +11,8 @@
 #' @param info String with information on the transformation. Used for the
 #'        print-method. Usually, this argument can be ignored
 #'
-#' @inheritParams esc_d2or
+#' @inheritParams convert_d2r
+#' @inheritParams convert_d2or
 #'
 #' @return The effect size \code{es}, the standard error \code{se}, the variance
 #'         of the effect size \code{var}, the lower and upper confidence limits
@@ -21,35 +22,40 @@
 #' @references Cohen J. 1988. Statistical Power Analysis for the Behavioral Sciences. 2nd ed. Hillsdale, NJ: Erlbaum
 #'
 #' @examples
-#' # d to f
-#' esc_d2f(d = 0.2, se = .1, totaln = 50)
+#' # d to eta squared
+#' convert_d2etasq(d = 0.7, se = 0.5, grp1n = 70, grp2n = 80)
 #'
 #' @export
-esc_d2f <- function(d, se, v, totaln, info = NULL, study = NULL) {
-  # compute effect size f
-  es <- d / 2
+convert_d2etasq <- function(d, se, v, grp1n, grp2n, info = NULL, study = NULL) {
 
   # check if parameter are complete
   if ((missing(se) || is.null(se) || is.na(se)) && (missing(v) || is.null(v) || is.na(v))) {
     warning("Either `se` or `v` must be specified.", call. = F)
-    return(esc_generic(es = NA, v = NA, es.type = "f", grp1n = NA, grp2n = NA, info = NA, study = NA))
+    return(esc_generic(es = NA, v = NA, es.type = "eta", grp1n = NA, grp2n = NA, info = NA, study = NA))
   }
 
   # do we have se?
-  if (!missing(se) && !is.null(se) && !is.na(se)) v <- se ^ 2
+  if (!missing(se) && !is.null(se) && !is.na(se)) v <- se^2
+
+  # do we have a separate info string?
+  if (is.null(info)) info <- "effect size d to effect size eta squared"
+
+  p <- grp1n / (grp1n + grp2n)
+  es <- eta_squared(d = d)
+  v <- v / (v + 1 / (p * (1 - p)))
 
   # return effect size f
   structure(
-    class = c("esc", "esc_d2f"),
+    class = c("esc", "convert_d2etasq"),
     list(
       es = es,
       se = sqrt(v),
       var = v,
-      ci.lo = lower_d(es, v),
-      ci.hi = upper_d(es, v),
+      ci.lo = exp(lower_d(es, v)),
+      ci.hi = exp(upper_d(es, v)),
       w = 1 / v,
-      totaln = totaln,
-      measure = "f",
+      totaln = grp1n + grp2n,
+      measure = "etasq",
       info = info,
       study = study
     )